requirements.txt
django
xlrd
django-push-notifications
